+++ src/editor/components/scene/list.tsx
@@ -77,6 +77,13 @@ export const SceneList: FC<{
 
   const onDelete = useCallback(async () => {
     if (!selector.selected) return
+
+    // eslint-disable-next-line no-alert
+    const confirmDeletion = confirm(
+      `Are you sure you want to delete "${selector.selected.definition.entity}"?`,
+    )
+    if (!confirmDeletion) return
+
     const id = selector.selected.uid
     history.record({
       type: 'delete',
